src/main/java/com/mookShopping/mook/repository/OrderRepository.java
package com.mookShopping.mook.repository;

import com.mookShopping.mook.domain.Order;
import jakarta.persistence.EntityManager;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Repository;

@Repository
public class OrderRepository {

    @Autowired
    EntityManager em;

    public Long saveOrder(Order order) {
        em.persist(order);
        return order.getId();
    }

    public Order findOne(Long id) {
        return em.find(Order.class, id);
    }
}
